Issue #690 - Post extensions follow-ups
