+++ modules/shared/internal/calculate-backoff-delay.ts
@@ -1,0 +1,10 @@
+export function calculateBackoffDelay(
+  attempt: number,
+  baseDelayMs: number,
+  maxDelayMs: number
+): number {
+  const exponentialDelay = baseDelayMs * Math.pow(2, attempt);
+  const cappedDelay = Math.min(exponentialDelay, maxDelayMs);
+  const jitter = 0.75 + Math.random() * 0.5; // 0.75 to 1.25
+  return Math.floor(cappedDelay * jitter);
+}
