+++ modules/shared/internal/normalize-flag.ts
@@ -1,0 +1,12 @@
+export function normalizeFlag(value: unknown, defaultValue: boolean): boolean {
+  if (value === null || value === undefined) return defaultValue;
+  if (typeof value === 'boolean') return value;
+  if (typeof value === 'number') return Boolean(value);
+  if (typeof value === 'string') {
+    const normalized = value.trim().toLowerCase();
+    if (['true', '1', 'yes', 'y', 'on'].includes(normalized)) return true;
+    if (['false', '0', 'no', 'n', 'off'].includes(normalized)) return false;
+    return defaultValue;
+  }
+  return Boolean(value);
+}
