+++ modules/shared/internal/read-trimmed-string-property.ts
@@ -1,0 +1,6 @@
+export function readTrimmedStringProperty(record: unknown, key: string): string | undefined {
+  const value = (record as any)?.[key];
+  if (typeof value !== 'string') return undefined;
+  const trimmed = value.trim();
+  return trimmed ? trimmed : undefined;
+}
